add pressure graph, rearrange grids, fix a few bugs
Most of the code is implementing the upper right hand corner
graph with straight D3 since Dimple.js was getting in the way.

It looks cool, but it's not quite right. Something to look into later.

Swapped the AI board with player2 so human players are on the
left & right.

Removed some dead code.

Moved the centering code into the websocket handler so it can also
cover the upper right hand corner graph.

Changed the colors a little for the background gradient.